speech/requirements.txt
google-cloud-translate==3.15.5
google-cloud-texttospeech==2.16.3
pydub==0.25.1
simpleaudio==1.0.4
numpy==1.26.4
